requirements.txt
Django==5.0
django-compat
django-cors-headers
djangorestframework
django-filter
mysqlclient==2.2.4
pymysql
django-rest-auth
cryptography
python-dateutil
google-cloud-storage
